requirements.txt
cement==2.10.2
paramiko==2.4.1
Jinja2==2.9.6
elasticsearch==6.2.0
pick==0.6.3
python_dateutil==2.7.3
docker
PyYAML==3.12
